Further optimizations to cable drawing and macro updates.

+++ src/ConvergeNode.cpp
@@ -14,7 +14,7 @@ void ConvergeNode::process() {
       float sum = 0.0f;
       for (const auto &n : chord)
         sum += n.noteNumber;
-      return sum / chord.size();
+      return sum / (float)chord.size();
     };
 
     std::stable_sort(steps.begin(), steps.end(),
@@ -27,11 +27,11 @@ void ConvergeNode::process() {
     int s = 0;
     int e = (int)steps.size() - 1;
     while (s <= e) {
-      if (!steps[s].empty()) {
-        sortedSeq.push_back(steps[s]);
+      if (!steps[(size_t)s].empty()) {
+        sortedSeq.push_back(steps[(size_t)s]);
       }
-      if (s != e && !steps[e].empty()) {
-        sortedSeq.push_back(steps[e]);
+      if (s != e && !steps[(size_t)e].empty()) {
+        sortedSeq.push_back(steps[(size_t)e]);
       }
       s++;
       e--;
